+++ src/ui/labels/footer_label.py
@@ -1,11 +1,11 @@
 """
-Componente para rodapé.
+Componente de rodapé.
 """
 
-from kivy.uix.label import Label
+from src.ui.labels.base_label import BaseLabel
 
 
-class FooterLabel(Label):
+class FooterLabel(BaseLabel):
     """
     Label utilizado para informações de rodapé.
     """
@@ -14,13 +14,6 @@ class FooterLabel(Label):
         super().__init__(
             text=text,
             font_size=14,
-            halign="center",
-            valign="middle",
             **kwargs,
         )
-
-        self.bind(size=self._update_text_size)
-
-    def _update_text_size(self, *_args) -> None:
-        self.text_size = self.size